pyproject.toml
[project]
name = "pyComputeFarm"
version = "0.0.1"
description = "A simple Python based collection of tools to run a distributed compute farm"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "aiofiles>=22.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.2",
    "click>=8.1.3",
    "cryptography>=38.0.3",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[project.scripts]
rcf = "rcf:cli.cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
